[examples] remove LEDs from sleepy-demo (#182)
This removes LED support from the sleepy-demo example apps

Some boards have limited GPIO and, as a result, button 0 and LED 0
share the same pin. The same is true for button 1 and LED 1 on these
boards.

Instead of special-casing, the decision was made to just remove LEDs
from the apps.

Users will still be able to see button presses by monitoring the CLI
output of both apps.