Add static struct for otRadioIeInfo in PAL (#427)
Nodes would be unable to join an established thread network if the
leader and prospective child each were configured with
OT_TIME_SYNC=ON.

Rather than joining the existing network, prospective child nodes
would fail to decode the child id response packets and form their own
partitions. This itself is resolved from the time that the issue was
reported by updating to a develop/22q2 patch 2 version of the project,
but an issue remained with network time values staying unsynchronized.

This synchronization issue was caused by an uninitialized pointer
member of a static TxFrame object in the PAL code which set the time
synchronization data to be values read from garbage memory instead of
time values obtained from the platform.

This commit adds a fix for the improper time sync values by
initializing the ieInfo data in the frame object, and the nodes should
be able to join the network.