[github-actions] fix 'Build and push' step of docker workflow
A bug was introduced in openthread/ot-efr32#793 where I moved
`Dockerfile` to `docker/Dockerfile` but did not update both the
build steps in the docker workflow, causing [this failure](https://github.com/SiliconLabs/ot-efr32/actions/runs/8468125501/job/23200371772#step:13:126).

This fixes the bug by specifying the new location of the Dockerfile
in the `Build and push` step.