[github-actions] build docker image then run `arm-gcc` with it
This changes the relationship of the workflows so that the `arm-gcc` job
in `build.yml` is run using the docker image generated by the docker
`build-image` job.

Previously, the `arm-gcc` job would just use the `:latest` tag of
`siliconlabsinc/ot-efr32-dev` from DockerHub. This presents a problem
when there are changes to the bootstrap script which are required for
the build script to succeed. These changes would not be present in the
currently published `siliconlabsinc/ot-efr32-dev:latest` since that
image was build off of `SiliconLabs/ot-efr32:main`, not the current branch.