[slc] add `ot_efr32` version of `ot_headers`
As part of SBOM work, openthread headers were pulled out of SLC components
which referenced mostly Silabs code and consolidated into the `ot_headers`
component. A requirement for `ot_headers` was then added to affected
components.

This requirement causes ALL OpenThread headers to be added to any component
which previously only required a subset of the headers.

In the context of ot-efr32 builds, this causes SLC generation failures as the
CMake platform library definitions will now reference all openthread headers in
`ot_headers`, which may or may not match up (ex. deleted) with the version of
OpenThread checked out in the `openthread` submodule in ot-efr32.

This commit adds an ot-efr32 version of `ot_headers` which fixes the issue by
pulling in only required headers instead of all headers in the `simplicity_sdk`
version of `ot_headers`.